moved to single manager
